verification_code/ocr/requirements.txt
Django==2.1.3
Pillow==5.3.0
pytz==2018.7
tesseract==0.1.3
tesserocr==2.3.1
